Update readme
Getting back to this - assuming that I don't have any more crazy participants breaking electrodes or some crap like that, I'm going to have a lot of data files on my hands by the end of the semester (hopefully about 20? maybe 30?). While I've figured out a way to do this in excel, I think it's worthwhile to spend the several hours (maybe a few days at my pace) trying to code this in python so I can quite literally just press a button and have it do the work for me. I've been busy, and goodness knows I'll still be busy - but, I have been reminded of the fact that I will never be less busy than I am at this particular moment. SO - with that in mind, I think this time-saver of a program will help a ton, and then can be modified easily for other data file cleaning if I have to do other methods for this particular EEG program.